Fixed crash with alternate format numbers
This appears to have been crashing for numbers formatted such as 4,3
instead of 4.3. Presumably this is to do with different output from the
`DecimalFormat.format()` call compared to what `Double.parseDouble` was
expecting. That shouldn't be the case, but I had a crash report here
nonetheless. Also, this method is faster.