Add live Map to the top of the main activity
In preparation for drawing latest quakes on it. This is a little
problematic, but I'm not sure if it's because of my Android M device.
The map attributes in the XML simply don't work for the fragment (they
work for the MapView, but that requires passing activity lifecycle
events through to it), and the status bar is always shown - the
fragment doesn't expand upwards to fill it.