Supplying a Context to getLocation method of EarthquakeListViewModel
I couldn't find this documented anywhere, but George Mount answered on
StackOverflow that a 'context' variable is available - so that gets
passed in, and all the formatting is done within the model. I'd still
rather have simple getter, but I'd also rather not retain a reference to
a context within the model.