Fix for Geonet API returning most recent event when it hasn't been updated
We were supplying a timestamp for which we want events that have been
updated after that timestamp. The API was always returning the event
that matched that timestamp, as if we were supplying '>=' instead of
'>'.

The workaround suggested by kfenaugh of Geonet was to add an additional
0 to the milliseconds entry. This appears to work, as if we supply the
second-to-last events updated timestamp, we get the most recent event,
and if we supply the most recent event, we get 0 results.