Move Earthquake retrieval to a service
Now this is how Rx was meant to be used (except actually it needs some
lambdas but that can come later).
Not sure about the Dagger implementation here - seems like a whole lot
of overhead when we need to think about Realm's place in
threading/Activity lifecycles anyway. Will re-evaluate after/as I
write/update unit tests - Dagger 1 was better for tests anyway.