Stop injecting Realm instances
We now inject the RealmConfiguration into the Application at startup
and then set that as default. The idea of injecting a Realm doesn't
work because of it's threading restrictions. Take the
NetworkRunnerService for example; if it injects in its constructor, the
Realm is created on the wrong thread. If it injects in its handleIntent
method, then it can't use any other injected objects until then. If we
use a Lazy injection, then every time we query for it we get the same
Realm - even after closing it!
The benefit of retrieving your own Realm when you need (apart from it
being easier to figure out what thread it should be accessed from) is
it also makes it more obvious when it should be closed.