Fix list dividers being transparent
When in landscape we have remove the background from the list, which
resulted in the "dividers" which we had before no longer being there.
Suspect they never existed and were simply showing the background the
whole time. Now we have a divider defined, with the colour taken from
the "background colours" in the support library - found the one that
appears visually closest to how it was before.