Remove timezone from query time formatter
Locally for me this would format the query time with "-0400" (my
current timezone). On users devices, it was using their timezone -
"+1200". Geonet couldn't deal with that. Instead, we send the time in
UTC (i.e., +0000) and drop the offset. I've no idea how this didn't
come up in testing - I had people in NZ test the app and it worked!