Switch from Retrofit to OkHttp and manual requests
This way we can read the body to a string before we parse it. This is
important, as I've been getting a few unexpected parse failures - the
server sends back a 200 response and when I re-run the logged request
URL it works. So I need to log the body to figure out what's going on,
but I can't do that since the body is a 1-shot stream. So here we read
the body to a string and only then do we try to parse it.