Remove paging from delta updates
It appears we can no longer use the "sortBy" param to order results
coming back from the server. Luckily we sort client-side at query time,
so that wasn't crucial. Unfortunately that also means we can no longer
retrieve paged results - chunks of 50 at a time, from
least-recently-updated to most-recently-updated. So instead I've decided
to go with an arbitrary "up to 1000 most recent events". This will have
no discernible effect for most users of the app.
This changes the timestamp format to also remove any timezone or
milliseconds information, as the server no longer accept timestamps in
that format.
https://github.com/GeoNet/help/issues/15