Update Realm (0.87.4 -> 4.3.4)
Realm now uses a Gradle plugin rather than just existing as a
standalone library. The callbacks also now return the changed object
(instead of nothing).

This code is pretty weird - it re-queries for the RealmObject in some
places, but not in others. Clearly whenever I wrote this I was either
confused, or not all-in on Realm, since:

> RealmObjects are live, auto-updating views into the underlying data;
you never have to refresh objects.

¯\_(ツ)_/¯