Introduce SyncCoordinator for handling download + storage during a sync
Incomplete - the test suite gives an example of a case where we expect
to store the output even when an error is emitted (and we expect to see
that error emission).

Ideally the buffer() operator would emit anything buffered before
propagating the error, however it explicitly does not do this.
Investigating a fix (possibly in Rx2?) - laptop had a kernel panic
though so I'm pushing this up in case of disaster.